move bind type to its own file
and remove Getter pattern
